requirements.txt
fastapi
uvicorn
pydantic
openai
langchain
langchain_community
langgraph
pymupdf
arxiv
python-dotenv
